+++ .gitignore
@@ -1,0 +1,2 @@
+#funciones/__pycache__
+**/__pycache__
